Generic Binding improvement - Fix .psd1, tests
