+++ src/gen/AzPowerShellSdkStrings.cs
@@ -47,7 +47,7 @@ internal class AzPowerShellSdkStrings {
     
     /// <summary>
     /// Looks up a localized string similar to 
-    ///   Could not add additional information with name {name} to binding {bindingName} because no binding with this name was found
+    ///   Could not add additional information with name {0} to binding {1} because no binding with this name was found
     /// </summary>
     internal static string AdditionalInformationNoSuchBinding {
         get {
@@ -97,21 +97,21 @@ internal class AzPowerShellSdkStrings {
 
     /// <summary>
     /// Looks up a localized string similar to 
-    ///   The following problems exist with GenericBinding: \n
-    /// </summary>
-    internal static string GenericBindingProblemsExist {
-        get {
-            return ResourceManager.GetString("GenericBindingProblemsExist", resourceCulture);
-        }
-    }
-
-    /// <summary>
-    /// Looks up a localized string similar to 
     ///   This function app directory contains functions which rely on host indexing, please remove them or configure this app for host indexing
     /// </summary>
     internal static string HybridModelDetected {
         get {
             return ResourceManager.GetString("HybridModelDetected", resourceCulture);
+        }
+    }
+
+    /// <summary>
+    /// Looks up a localized string similar to 
+    ///   The following problems exist with InputBinding: \n
+    /// </summary>
+    internal static string InputBindingProblemsExist {
+        get {
+            return ResourceManager.GetString("InputBindingProblemsExist", resourceCulture);
         }
     }
 
@@ -192,6 +192,16 @@ internal class AzPowerShellSdkStrings {
     internal static string MissingValue {
         get {
             return ResourceManager.GetString("MissingValue", resourceCulture);
+        }
+    }
+
+    /// <summary>
+    /// Looks up a localized string similar to 
+    ///   The following problems exist with OutputBinding: \n
+    /// </summary>
+    internal static string OutputBindingProblemsExist {
+        get {
+            return ResourceManager.GetString("OutputBindingProblemsExist", resourceCulture);
         }
     }
